requirements.txt
faiss-cpu==1.8.0
pandas==2.2.3
scikit_learn==1.6.1
sentence_transformers==3.3.1
termcolor==2.5.0
torch==2.3.0
tqdm==4.66.4
transformers==4.44.2
nltk>=3.8.1
matplotlib>=3.8.0
